userList to each identifier

+++ src/components/choiceRoomWindow/JoinRoom.tsx
@@ -1,27 +1,40 @@
 import React, {useState} from "react";
 import style from './choiceRoomWindow.module.scss';
-import {useTypeDispatch} from "../../store/store";
+import {RootState, useTypeDispatch} from "../../store/store";
 import {useNavigate} from "react-router-dom";
 import {setAdmin} from "../../store/web-slices/role_slice";
 import {ROOM_ID_IN_STORAGE} from "../../store/web-slices/chat_slice";
-import {setAuth} from "../../store/web-slices/profile_slice";
+import {checkExistingRoom, setAuth} from "../../store/web-slices/profile_slice";
+import {useSelector} from "react-redux";
 
 
 function JoinRoom() {
     const [idRoom, setIdRoom] = useState('');
     const dispatch = useTypeDispatch();
+    const {isRoomExist} = useSelector((state : RootState) => state.profileReducer)
+    const [isErrorRoom, setIsErrorRoom] = useState(false)
 
+    //Баг со входом хз надо как то эвейт на диспатче правильно поставить чтобы он обновлялся норм
     let navigate = useNavigate();
-    const handleJoinRoom = () => {
-        sessionStorage.setItem(ROOM_ID_IN_STORAGE, idRoom);
-        dispatch(setAdmin(false))
-        dispatch(setAuth(true));
-        navigate('/enter');
+    const handleJoinRoom = async () => {
+        await dispatch(checkExistingRoom(idRoom))
+        console.log(isRoomExist)
+        if (isRoomExist) {
+            sessionStorage.setItem(ROOM_ID_IN_STORAGE, idRoom);
+            dispatch(setAdmin(false))
+            dispatch(setAuth(true));
+            navigate('/enter');
+            setIsErrorRoom(false)
+        }
+        else {
+            setIsErrorRoom(true)
+        }
+
     }
 
-    const handlePressEnter = (e: React.KeyboardEvent<HTMLInputElement>) => {
+    const handlePressEnter = async (e: React.KeyboardEvent<HTMLInputElement>) => {
         if (e.key === "Enter")
-            handleJoinRoom();
+            await handleJoinRoom();
     }
 
     return (
@@ -29,10 +42,14 @@ function JoinRoom() {
             <div className={style.widget}>
                 <>
                     <div>Индефикатор комнаты:</div>
+                    {isErrorRoom && <div className={style.errorMessage}>Такой комнаты не существует</div>}
                     <input className={`${style.input} input`}
                            type="text"
                            value={idRoom}
-                           onChange={e => setIdRoom(e.target.value)}
+                           onChange={e => {
+                               setIdRoom(e.target.value)
+                               setIsErrorRoom(false)
+                           }}
                            onKeyPress={e => handlePressEnter(e)}
                     />
                 </>
